nav overhaul: move groupPins into gropus state
